requirements.txt
python-telegram-bot
pycricbuzz